modification des topics pour utiliser ResponseEntity

+++ src/main/java/dev/controller/TopicCtrl.java
@@ -1,12 +1,18 @@
 package dev.controller;
 
+import dev.controller.dto.ErrorDTO;
 import dev.controller.dto.TopicDTO;
 import dev.controller.dto.TopicModifDTO;
 import dev.entite.forum.Topic;
 import dev.service.TopicService;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @CrossOrigin
 @RestController
@@ -19,20 +25,36 @@ public class TopicCtrl {
     }
 
     @GetMapping
-    // A FAIRE. Utilisé un ResponseEntity
-    public List<Topic> getAllTopics() {
-        return topicService.findAll();
+    public ResponseEntity<?> getAllTopics() {
+        List<Topic> topics = topicService.findAll();
+        if(!topics.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.OK).body(topics);
+        } else {
+            return ResponseEntity.status(HttpStatus.NO_CONTENT)
+                    .body("Il n'y a aucun topic d'enregistré");
+        }
     }
 
     @PostMapping
-    // A FAIRE. Utilisé un ResponseEntity
-    public Topic create(@RequestBody TopicDTO topicDTO) {
-        return topicService.create(topicDTO);
+    public ResponseEntity<?> create(@RequestBody @Validated TopicDTO topicDTO, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                    .body(new ErrorDTO(List.of("Les informations saisies ne sont pas valides")));
+        } else {
+            return ResponseEntity.status(HttpStatus.CREATED)
+                    .body(topicService.create(topicDTO));
+        }
     }
+
     @DeleteMapping("/{id}")
-    // A FAIRE. Utilisé un ResponseEntity
-    public void delete(@PathVariable int id ){
-        topicService.delete(id);
+    public ResponseEntity<?> delete(@PathVariable int id) {
+        Optional<Topic> topic = topicService.findById(id);
+        if (topic.isPresent()) {
+            topicService.delete(id);
+            return ResponseEntity.status(HttpStatus.OK).build();
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
     }
 
     @PatchMapping
